pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Joichiro433"]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
numpy = ">=1.22,<1.23"
pandas = "^1.4.2"
matplotlib = "^3.5.2"
seaborn = "^0.11.2"
scikit-learn = "^1.1.1"
nptyping = "^2.1.1"
# tensorflow = "^2.9.1"
tensorflow-macos = "^2.9.2"
pydot = "^1.4.2"
graphviz = "^0.20"
rich = "^12.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
